move events view matcher + v1 filter compiler from rpcv2 into shared/store
